refactor(#310c): extract chain_job to app.tasks._chain (break import cycle)
CodeQL flagged three "Cyclic import" alerts on PR #1050 (#644/#645/
#646) that all clustered to one root cause:

    bootstrap.py     (lazy in bootstrap_handlers) → ingestion_tasks
    ingestion_tasks  (top-level)                  → runner
    runner.py        (lazy in run_job)            → bootstrap

Runtime was safe (back-edges live inside function bodies, so the
import re-entry never happened in practice), but the static graph
was cyclic and CodeQL is correct to flag it — cycles complicate
onboarding and signal that the module boundaries are wrong.

Fix: move ``chain_job`` (the only symbol ``ingestion_tasks`` imports
from ``runner``) into a dedicated ``app.tasks._chain`` module.  The
top-level ``ingestion_tasks → runner`` edge becomes ``ingestion_tasks
→ _chain``; ``_chain`` only needs ``run_job`` for the
``fire_and_forget`` dispatch, which is a function-body call resolved
via lazy import.  The static module-load graph is now acyclic and
CodeQL's ``py/cyclic-import`` rule clears all three alerts without
any ``# codeql[…]`` suppression.

Adjacent moves:
- Tests for ``chain_job`` move from ``test_runner.py`` to a sibling
  ``test_chain.py`` mirroring the production-side split.  Coverage
  unchanged (3 tests: pipeline_id inheritance, generate-when-missing,
  explicit overrides).
- Two integration tests that patched ``app.tasks.runner.fire_and_forget``
  move the patch path to ``app.tasks._chain.fire_and_forget`` since
  that's where the dispatch call now lives.
- Two docstring references to ``runner.chain_job`` updated for accuracy.

A static-import regression test in
``test_defer_finalize.py::test_chain_job_no_longer_importable_from_runner``
(landed with the previous commit) pins the new location: a future
refactor that puts ``chain_job`` back in runner.py would re-open the
cycle and fail this test loudly.
